10_21

+++ model/res.go
@@ -2,6 +2,6 @@ package model
 
 // Response .
 type Response struct {
-	Code 	uint32	`json:"code"`
+	Code 	uint8	`json:"code"`
 	Msg 	string 	`json:"msg"`
 }
